# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EDUCAFI",
        "description": "Documentación de la API EDUCAFI",
        "contact": {
            "email": "soporte@educafi.mx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/alumnos/{numero_cuenta}": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Mostrar un alumno por número de cuenta",
                "operationId": "95a4765efe8b593bae6642417d82a646",
                "parameters": [
                    {
                        "name": "numero_cuenta",
                        "in": "path",
                        "description": "Número de cuenta del alumno",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del alumno"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                }
            }
        },
        "/api/alumnos/buscar": {
            "post": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Buscar alumnos por criterios",
                "operationId": "8efbea37da112d6fc03160a506ffdf09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "correo_personal": {
                                        "type": "string",
                                        "example": "juan@mail.com"
                                    },
                                    "numero_cuenta": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultados de la búsqueda"
                    },
                    "400": {
                        "description": "Debe enviar al menos un campo para buscar"
                    },
                    "500": {
                        "description": "Error al ejecutar la consulta"
                    }
                }
            }
        },
...
        "/APIeducafi/api/prebecarios/buscar": {
            "post": {
                "tags": [
                    "Prebecarios"
                ],
                "summary": "Buscar candidatos a prebecarios",
                "operationId": "125e9b63b01131806422a20f895cc3d6",
                "requestBody": {
                    "description": "Filtros opcionales para buscar candidatos a prebecarios",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "avance_min": {
                                        "description": "Porcentaje mínimo de avance",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "avance_max": {
                                        "description": "Porcentaje máximo de avance",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "carrera": {
                                        "description": "Nombre de la carrera",
                                        "type": "string",
                                        "example": "Ingeniería en Computación"
                                    },
                                    "promedio_min": {
                                        "description": "Promedio mínimo",
                                        "type": "number",
                                        "format": "float",
                                        "example": 8
                                    },
                                    "promedio_max": {
                                        "description": "Promedio máximo",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "semestre_ingreso": {
                                        "description": "Semestre de ingreso (formato: AAAA-N)",
                                        "type": "string",
                                        "example": "2023-2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de candidatos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "nombre": "Ana Torres López",
                                            "carrera": "Ingeniería en Computación",
                                            "promedio": 9.2,
                                            "avance": 75,
                                            "semestre_ingreso": "2022-1"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}